courses/genaiboot/requirements.txt
# Core Requirements for Generative AI Bootcamp

# OpenAI
openai>=1.12.0
tiktoken>=0.5.1

# Anthropic
anthropic>=0.18.0

# HuggingFace
transformers>=4.36.0
torch>=2.1.0
datasets>=2.14.0
accelerate>=0.24.0

# LangChain & RAG
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.0.2
llama-index>=0.9.0
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Vector Databases
pinecone-client>=2.2.0
weaviate-client>=3.24.0
qdrant-client>=1.6.0
pgvector>=0.2.0

# Agents & Workflows
langgraph>=0.0.20
crewai>=0.16.0

# Web & API
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.29.0
gradio>=4.7.0
flask>=3.0.0
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# AWS & Cloud
boto3>=1.28.0
sagemaker>=2.190.0
awscli>=1.29.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Development Tools
python-dotenv>=1.0.0
jupyter>=1.0.0
ipython>=8.16.0
pytest>=7.4.0
black>=23.10.0
ruff>=0.1.0
mypy>=1.6.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
pymongo>=4.5.0

# Audio/Image Processing
Pillow>=10.1.0
opencv-python>=4.8.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1

# Utilities
pydantic>=2.4.0
tqdm>=4.66.0
rich>=13.6.0
click>=8.1.0
typer>=0.9.0
pyyaml>=6.0.1
jsonlines>=4.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0
sphinx>=7.2.0
